+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Issue #160: Add motor theta configuration
+- Issue #162: Add sensor receptive angle in radar
 
 ### Changed